refactor: LottoResult의 당첨횟수 단위를 Long->Int로 변경

+++ src/main/java/lotto/model/LottoStatistics.java
@@ -7,35 +7,35 @@ import java.util.List;
 import java.util.Map;
 
 public class LottoStatistics {
-    private final Map<LottoResult, Long> counts;
+    private final Map<LottoResult, Integer> counts;
     private final PurchaseAmount purchaseAmount;
 
-    private LottoStatistics(Map<LottoResult, Long> counts, PurchaseAmount purchaseAmount) {
+    private LottoStatistics(Map<LottoResult, Integer> counts, PurchaseAmount purchaseAmount) {
         this.counts = counts;
         this.purchaseAmount = purchaseAmount;
     }
 
     public static LottoStatistics from(List<LottoResult> results, PurchaseAmount purchaseAmount) {
-        Map<LottoResult, Long> counts = initializeCounts();
-        results.forEach(result -> counts.put(result, counts.get(result) + 1L));
+        Map<LottoResult, Integer> counts = initializeCounts();
+        results.forEach(result -> counts.put(result, counts.get(result) + 1));
         return new LottoStatistics(counts, purchaseAmount);
     }
 
-    private static Map<LottoResult, Long> initializeCounts() {
-        Map<LottoResult, Long> counts = new EnumMap<>(LottoResult.class);
+    private static Map<LottoResult, Integer> initializeCounts() {
+        Map<LottoResult, Integer> counts = new EnumMap<>(LottoResult.class);
         for (LottoResult result : LottoResult.values()) {
-            counts.put(result, 0L);
+            counts.put(result, 0);
         }
         return counts;
     }
 
-    public long countOf(LottoResult lottoResult) {
+    public int countOf(LottoResult lottoResult) {
         return counts.get(lottoResult);
     }
 
-    public long calculateTotalPrize() {
+    public int calculateTotalPrize() {
         return counts.entrySet().stream()
-                .mapToLong(entry -> entry.getKey().getPrize() * entry.getValue())
+                .mapToInt(entry -> entry.getKey().getPrize() * entry.getValue())
                 .sum();
     }
